Expose leaf ctree nodes in Query, GetLeaf, Walk
This change is for cache, which puts leaves into a coalescing queue. Currently
it wraps values with a special type with mutex, in case those leaves get
updated while client is still receiving previous nodes.

After this change we should be able to get rid of the Node wrapper with its
extra mutex, and reuse the mutex inside ctree nodes.

PiperOrigin-RevId: 173296027