Fix targets relying on unspecified or buggy Merge/Clone behavior
The proto.Merge function is semantically equivalent to:
  func Merge(dst, src proto.Message) {
    b, _ := Marshal(src)
    UnmarshalMerge(b, dst)
  }

Any usages of Merge that rely on artifacts that cannot be represented
by the protobuf binary format are depending on undefined behavior
that so happened to work. See b/141414724 for examples of such cases.

We fix usages of Merge/Clone that we detected via TAP global presubmit
to be broken by an implementation of Merge that strictly obeys the
protobuf data model.

The fix involves either:
* modifying the test to be less brittle to minor changes to Merge semantics.
* modifying the code to have a custom version of Clone/Merge that preserves
the prior semantics.

Tested:
    TAP --sample ran all affected tests and none failed
    http://test/OCL:270510284:BASE:270501421:1569137215080:eac3b8b8
PiperOrigin-RevId: 270688809