Fix targets to avoid unstable string comparisons
The Message.String method is intended for debugging and does not guarantee
output stability. Code that relies on the output being stable for all time is
brittle and has unintentionally become a change detector for the proto package.
The v2 implementation of prototext (which will become the underlying
implementation of Message.String) will have deliberately unstable output
to prevent future code from relying on this unspecified property.

If the test code is changed, it was probably fixed by parsing the hard-coded
text format and reformatting it at runtime with the current implementation
of Message.String. This may not be the most elegant fix, but is easy to
mechanically perform and produces minimal changes in the user's code.
See http://go/go-test-comments#compare-stable-results

If the non-test code is changed, it was either because:
  1) The tests depending on String method stability were either
  too difficult or too many to change; in this case it was easier to
  modify the user code to preserve the same behavior. OR
  2) The proper functioning of the non-test code actually depends on
  the stability of the String method (e.g., using the String as a key
  into a map or database).
In both cases, the code was modified to use proto.CompactTextString,
which was the old underlying implementation of Message.String.

PiperOrigin-RevId: 273338794